requirements.txt
Flask==0.10.1
Flask-MySQLdb==0.2.0
gunicorn==18.0
itsdangerous==0.24
Jinja2==2.8
MarkupSafe==0.23
MySQL-python==1.2.5
mysqlclient==1.3.6
requests==2.8.1
Werkzeug==0.10.4
wheel==0.24.0
